fix: 코드리뷰 관련 사항 수정
