+++ package.json
@@ -34,7 +34,6 @@
     "clsx": "^1.2.1",
     "framer-motion": "^6.5.1",
     "graphql": "^16.5.0",
-    "lodash": "^4.17.21",
     "next": "^12.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -61,7 +60,6 @@
     "@svgr/webpack": "^6.3.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
-    "@types/lodash": "^4.14.182",
     "@types/node": "^18.0.6",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
